+++ src/config.rs
@@ -14,6 +14,7 @@ use std::path::PathBuf;
 pub struct Config {
     pub path: PathBuf,
     pub display_mode: DisplayMode,
+    pub git_path: Option<PathBuf>,
 }
 
 // "EntryConfig" is a reflection of "Config" with its fields wrapped as "Option" types. This is to
@@ -25,6 +26,7 @@ pub struct Config {
 struct EntryConfig {
     pub path: Option<PathBuf>,
     pub display_mode: Option<DisplayMode>,
+    pub git_path: Option<PathBuf>,
 }
 
 // "DisplayMode" dictates which way the results gathered should be displayed to the user via
@@ -70,7 +72,7 @@ impl Config {
     }
 
     // This method does not look for the config file and uses "EntryConfig"'s defaults instead.
-    // This method is best for testing use and when the user wishes to skip config file lookup.
+    // It is best for testing use and when the user wishes to skip config file lookup.
     pub fn new() -> Result<Config> {
         entry_config_to_config(&EntryConfig::default())
     }
@@ -93,5 +95,6 @@ fn entry_config_to_config(entry_config: &EntryConfig) -> Result<Config> {
             Some(s) => s.clone(),
             None => DisplayMode::Standard,
         },
+        git_path: entry_config.git_path.clone(),
     })
 }